Move call device handling out of WebRTCSession
