+++ src/CallManager.h
@@ -8,6 +8,7 @@
 #include <QString>
 #include <QTimer>
 
+#include "CallDevices.h"
 #include "WebRTCSession.h"
 #include "mtx/events/collections.hpp"
 #include "mtx/events/voip.hpp"
@@ -53,7 +54,7 @@ public:
 public slots:
         void sendInvite(const QString &roomid, bool isVideo);
         void syncEvent(const mtx::events::collections::TimelineEvents &event);
-        void refreshDevices() { session_.refreshDevices(); }
+        void refreshDevices() { CallDevices::instance().refresh(); }
         void toggleMicMute();
         void toggleCameraView() { session_.toggleCameraView(); }
         void acceptInvite();
